model-backend/nirvik/requirements.txt
fastapi
python-multipart
uvicorn
opencv-python-headless
numpy
tensorflow-cpu